PeripheralのName取得処理をproxyへ移動
